break out getMatchesForUrl from createResolvedRouteForUrl
