+++ src/services/createResolvedRouteForUrl.browser.spec.ts
@@ -1,169 +1,7 @@
-import { afterEach, expect, test, vi } from 'vitest'
+import { expect, test } from 'vitest'
 import { createRoute } from '@/services/createRoute'
 import { createResolvedRouteForUrl } from '@/services/createResolvedRouteForUrl'
-import * as utilities from '@/services/routeMatchScore'
-import { Route } from '@/types'
 import { component } from '@/utilities/testHelpers'
-
-afterEach(() => {
-  vi.restoreAllMocks()
-})
-
-test('given path WITHOUT params, returns match', () => {
-  const parent = createRoute({
-    name: 'parent',
-    path: '/parent',
-  })
-
-  const child = createRoute({
-    parent: parent,
-    name: 'parent.child',
-    path: '/child',
-  })
-
-  const grandchild = createRoute({
-    parent: child,
-    name: 'parent.child.grandchild',
-    path: '/grandchild',
-    component,
-  })
-
-  const routes = [parent, child, grandchild]
-
-  const match = createResolvedRouteForUrl(routes, '/parent/child/grandchild')
-
-  expect(match?.name).toBe('parent.child.grandchild')
-  expect(match?.matched.name).toBe('parent.child.grandchild')
-})
-
-test('given path to unnamed parent, without option to get to leaf, returns undefined', () => {
-  const unnamedParent = createRoute({
-    path: '/unnamed',
-  })
-
-  const unnamedChild = createRoute({
-    parent: unnamedParent,
-    path: '/unnamed-child/[child-id]',
-  })
-
-  const namedGrandchild = createRoute({
-    parent: unnamedChild,
-    path: '/named-grandchild',
-    component,
-  })
-
-  const routes = [unnamedParent, unnamedChild, namedGrandchild]
-
-  const match = createResolvedRouteForUrl(routes, '/unnamed')
-
-  expect(match).toBeUndefined()
-})
-
-test('given path to unnamed  parent, with option to get to leaf, returns available leaf', () => {
-  const unnamedParent = createRoute({
-    path: '/unnamed',
-  })
-
-  const unnamedChildRoot = createRoute({
-    parent: unnamedParent,
-    name: 'child-root',
-    component,
-  })
-
-  const routes = [unnamedChildRoot, unnamedParent]
-  const match = createResolvedRouteForUrl(routes, '/unnamed')
-
-  expect(match?.name).toBe('child-root')
-})
-
-test('given path that includes named parent and path to leaf, return first match', () => {
-  const namedParent = createRoute({
-    name: 'namedParent',
-    path: '/named-parent',
-  })
-
-  const namedChild = createRoute({
-    parent: namedParent,
-    name: 'namedParent.namedChild',
-  })
-
-  const namedGrandchild = createRoute({
-    parent: namedChild,
-    name: 'namedParent.namedChild.namedGrandchild',
-    component,
-  })
-
-  const routes = [namedParent, namedChild, namedGrandchild]
-  const match = createResolvedRouteForUrl(routes, '/named-parent')
-
-  expect(match?.name).toBe('namedParent.namedChild.namedGrandchild')
-  expect(match?.matched.name).toBe('namedParent.namedChild.namedGrandchild')
-})
-
-test('given route with simple string param WITHOUT value present, returns undefined', () => {
-  const route = createRoute({
-    name: 'simple-params',
-    path: '/simple/[simple]',
-    component,
-  })
-  const response = createResolvedRouteForUrl([route], '/simple/')
-
-  expect(response).toBeUndefined()
-})
-
-test('given route with simple string query param WITHOUT value present, returns undefined', () => {
-  const route = createRoute({
-    name: 'simple-params',
-    path: '/missing',
-    query: 'simple=[simple]',
-    component,
-  })
-  const response = createResolvedRouteForUrl([route], '/missing?without=params')
-
-  expect(response).toBeUndefined()
-})
-
-test('given route with equal matches, returns route with highest score', () => {
-  vi.spyOn(utilities, 'getRouteScoreSortMethod').mockImplementation(() => {
-    return (route: Route) => {
-      return route.name === 'second-route' ? -1 : +1
-    }
-  })
-
-  const routes = [
-    createRoute({
-      name: 'first-route',
-      path: '/',
-      component,
-    }),
-    createRoute({
-      name: 'second-route',
-      path: '/',
-      component,
-    }),
-    createRoute({
-      name: 'third-route',
-      path: '/',
-      component,
-    }),
-  ]
-
-  const response = createResolvedRouteForUrl(routes, '/')
-
-  expect(response?.name).toBe('second-route')
-})
-
-test('given a route without params or query returns an empty params and query', () => {
-  const route = createRoute({
-    name: 'route',
-    path: '/',
-    component,
-  })
-  const response = createResolvedRouteForUrl([route], '/')
-
-  expect(response?.params).toMatchObject({})
-  expect(response?.query).toMatchObject({})
-})
 
 test('given a url with a query returns all query values', () => {
   const route = createRoute({
