+++ src/services/createResolvedRouteForUrl.ts
@@ -1,42 +1,28 @@
-import { getRouteParamValues, routeParamsAreValid } from '@/services/paramValidation'
-import { isNamedRoute, routePathMatches, routeQueryMatches, routeHashMatches } from '@/services/routeMatchRules'
-import { getRouteScoreSortMethod } from '@/services/routeMatchScore'
+import { getRouteParamValues } from '@/services/paramValidation'
 import { ResolvedRoute } from '@/types/resolved'
 import { Routes } from '@/types/route'
-import { RouteMatchRule } from '@/types/routeMatchRule'
 import { parseUrl } from './urlParser'
 import { createResolvedRouteQuery } from './createResolvedRouteQuery'
 import { getStateValues } from './state'
 import { asUrl } from '@/types'
-
-const rules: RouteMatchRule[] = [
-  isNamedRoute,
-  routePathMatches,
-  routeQueryMatches,
-  routeHashMatches,
-  routeParamsAreValid,
-]
+import { getMatchesForUrl } from './getMatchesForUrl'
 
 export function createResolvedRouteForUrl(routes: Routes, url: string, state?: unknown): ResolvedRoute | undefined {
-  const sortByRouteScore = getRouteScoreSortMethod(url)
-
-  const matches = routes
-    .filter((route) => rules.every((test) => test(route, url)))
-    .sort(sortByRouteScore)
+  const matches = getMatchesForUrl(routes, url)
 
   if (matches.length === 0) {
     return undefined
   }
 
   const [route] = matches
-  const { search, hash } = parseUrl(url)
+  const { searchParams, hash } = parseUrl(url)
 
   return {
     id: route.id,
     matched: route.matched,
     matches: route.matches,
     name: route.name,
-    query: createResolvedRouteQuery(search),
+    query: createResolvedRouteQuery(searchParams),
     params: getRouteParamValues(route, url),
     state: getStateValues(route.state, state),
     hash,
